* 'master' of https://github.com/delacroix821026/p2p:
  1212121
  121212
  Config
  Excel Model

# Conflicts:
#	p2p_schedule/domain/src/main/java/com/newland/financial/p2p/domain/entity/ExcelOrderModel.java
#	p2p_schedule/service/src/main/java/com/newland/financial/p2p/service/impl/TransToSmallLoanConfig.java

+++ p2p_server/dao/src/main/java/com/newland/financial/p2p/dao/impl/DebitAndCreditDao.java
@@ -45,9 +45,12 @@ public class DebitAndCreditDao extends
      * @return boolean删除成功返回true, 失败false
      */
     public boolean deleteDebitAndCredit(final String userId) {
-        Map<String, Object> map = new HashMap<String, Object>();
-        map.put("userId", userId);
-        return super.deletes("deleteByProperties", map);
+        if (userId != null && userId.length() != 0) {
+            Map<String, Object> map = new HashMap<String, Object>();
+            map.put("userId", userId);
+            return super.deletes("deleteByProperties", map);
+        }
+        return false;
     }
 
     /**
